+++ kamehouse-commons-core/src/main/java/com/nicobrest/kamehouse/commons/model/systemcommand/MouseClickJvncSenderSystemCommand.java
@@ -28,12 +28,11 @@ public class MouseClickJvncSenderSystemCommand extends JvncSenderSystemCommand {
   public MouseClickJvncSenderSystemCommand(MouseButton mouseButton, int positionX, int positionY,
       int clickCount) {
     super();
-    logCommand = false;
-    executeOnDockerHost = true;
     this.mouseButton = mouseButton;
     this.positionX = positionX;
     this.positionY = positionY;
     this.clickCount = clickCount;
+    setDockerHostJvncSenderCommands();
   }
 
   /**
@@ -50,13 +49,12 @@ public class MouseClickJvncSenderSystemCommand extends JvncSenderSystemCommand {
   public MouseClickJvncSenderSystemCommand(MouseButton mouseButton, int positionX, int positionY,
       int clickCount, int sleepTime) {
     super();
-    logCommand = false;
-    executeOnDockerHost = true;
     setSleepTime(sleepTime);
     this.mouseButton = mouseButton;
     this.positionX = positionX;
     this.positionY = positionY;
     this.clickCount = clickCount;
+    setDockerHostJvncSenderCommands();
   }
 
   @Override
@@ -66,5 +64,11 @@ public class MouseClickJvncSenderSystemCommand extends JvncSenderSystemCommand {
     } catch (Exception e) {
       throw new KameHouseException(e);
     }
+  }
+
+  @Override
+  protected String buildKameHouseCmdJvncSenderOperationArgs() {
+    return "-mouseClick " + mouseButton.name() + "," + positionX + "," + positionY + ","
+        + clickCount;
   }
 }