+++ kamehouse-shell/bin/kamehouse/kamehouse-mobile-upload-apk-to-device.sh
@@ -13,7 +13,6 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseCfg
 loadKamehouseShellPwd
 
 SKIP_BUILD_MOBILE=false